feat: Rewrite api to typescript, update storyblok-js-client, create core custom base class
