--- src/api/config.js
@@ -1,25 +1,0 @@
-const StoryblokClient = require("storyblok-js-client")
-const {
-  oauthToken,
-  githubToken,
-  accessToken,
-  storyblokApiUrl
-} = require("../config/config")
-
-const headers = {
-  "Content-Type": "application/json",
-  Authorization: oauthToken
-}
-
-const githubHeaders = {
-  ...headers,
-  Authorization: `token ${githubToken}`
-}
-
-const sbApi = new StoryblokClient({ accessToken, oauthToken }, storyblokApiUrl)
-
-module.exports = {
-  headers,
-  githubHeaders,
-  sbApi
-}
