--- src/api/mutateComponents.js
@@ -1,57 +1,0 @@
-const Logger = require("../utils/loggerumd")
-const { spaceId } = require("../config/config")
-const { sbApi } = require("./config")
-const _resolvePresets = require("./resolvePresets")
-
-// UPDATE
-const updateComponent = (component, presets) => {
-  Logger.log(`Trying to update '${component.name}' with id ${component.id}`)
-  const componentWithPresets = component
-  const { all_presets, ...componentWithoutPresets } = componentWithPresets
-
-  sbApi
-    .put(`spaces/${spaceId}/components/${component.id}`, {
-      component: componentWithoutPresets
-    })
-    .then(res => {
-      Logger.success(`Component '${component.name}' has been updated.`)
-      if (presets) {
-        _resolvePresets(res, all_presets, component)
-      }
-    })
-    .catch(err => {
-      Logger.error("error happened... :(")
-      console.log(
-        `${err.message} in migration of ${component.name} in updateComponent function`
-      )
-    })
-}
-
-// CREATE
-const createComponent = (component, presets) => {
-  Logger.log(`Trying to create '${component.name}'`)
-  const componentWithPresets = component
-  const { all_presets, ...componentWithoutPresets } = componentWithPresets
-
-  sbApi
-    .post(`spaces/${spaceId}/components/`, {
-      component: componentWithoutPresets
-    })
-    .then(res => {
-      Logger.success(`Component '${component.name}' has been created.`)
-      if (presets) {
-        _resolvePresets(res, all_presets, component)
-      }
-    })
-    .catch(err => {
-      Logger.error("error happened... :(")
-      console.log(
-        `${err.message} in migration of ${component.name} in createComponent function`
-      )
-    })
-}
-
-module.exports = {
-  updateComponent,
-  createComponent
-}
