+++ src/commands/backup.ts
@@ -1,6 +1,7 @@
-import { Command, flags } from "@oclif/command";
+import { flags } from "@oclif/command";
+import Command from '../core'
 
-import { sbmigWorkingDirectory } from "../config/config";
+import storyblokConfig from "../config/config";
 import { getAllComponents, getComponent, getAllComponentsGroups, getComponentsGroup } from "../api/components";
 import { getAllDatasources, getDatasource, getDatasourceEntries } from '../api/datasources';
 import { getComponentPresets } from '../api/componentPresets';
@@ -37,10 +38,10 @@ export default class Backup extends Command {
         .then(async (res: any) => {
           const randomDatestamp = new Date().toJSON();
           const filename = `all-components-${randomDatestamp}`;
-          await createDir(`${sbmigWorkingDirectory}/components/`);
+          await createDir(`${storyblokConfig.sbmigWorkingDirectory}/components/`);
           await createJsonFile(
             JSON.stringify(res),
-            `${sbmigWorkingDirectory}/components/${filename}.json`
+            `${storyblokConfig.sbmigWorkingDirectory}/components/${filename}.json`
           );
           Logger.success(`All components written to a file:  ${filename}`);
           return true;
@@ -59,10 +60,10 @@ export default class Backup extends Command {
         if (res) {
           const randomDatestamp = new Date().toJSON()
           const filename = `component-${flags.oneComponent}-${randomDatestamp}`
-          await createDir(`${sbmigWorkingDirectory}/components/`)
+          await createDir(`${storyblokConfig.sbmigWorkingDirectory}/components/`)
           await createJsonFile(
             JSON.stringify(res),
-            `${sbmigWorkingDirectory}/components/${filename}.json`
+            `${storyblokConfig.sbmigWorkingDirectory}/components/${filename}.json`
           )
           Logger.success(
             `Component for ${flags.oneComponent} written to a file:  ${filename}`
@@ -79,10 +80,10 @@ export default class Backup extends Command {
       return getAllComponentsGroups().then(async res => {
         const randomDatestamp = new Date().toJSON()
         const filename = `all-component_groups-${randomDatestamp}`
-        await createDir(`${sbmigWorkingDirectory}/component_groups/`)
+        await createDir(`${storyblokConfig.sbmigWorkingDirectory}/component_groups/`)
         await createJsonFile(
           JSON.stringify(res),
-          `${sbmigWorkingDirectory}/component_groups/${filename}.json`
+          `${storyblokConfig.sbmigWorkingDirectory}/component_groups/${filename}.json`
         )
         Logger.success(`All groups written to a file:  ${filename}`)
       })
@@ -97,10 +98,10 @@ export default class Backup extends Command {
         if (res) {
           const randomDatestamp = new Date().toJSON()
           const filename = `components_group-${flags.oneComponentsGroup}-${randomDatestamp}`
-          await createDir(`${sbmigWorkingDirectory}/component_groups/`)
+          await createDir(`${storyblokConfig.sbmigWorkingDirectory}/component_groups/`)
           await createJsonFile(
             JSON.stringify(res),
-            `${sbmigWorkingDirectory}/component_groups/${filename}.json`
+            `${storyblokConfig.sbmigWorkingDirectory}/component_groups/${filename}.json`
           )
           Logger.success(
             `Components group for ${flags.oneComponentsGroup} written to a file:  ${filename}`
@@ -117,10 +118,10 @@ export default class Backup extends Command {
       return getAllDatasources().then(async (res: any) => {
         const randomDatestamp = new Date().toJSON()
         const filename = `all-datasources-${randomDatestamp}`
-        await createDir(`${sbmigWorkingDirectory}/datasources/`)
+        await createDir(`${storyblokConfig.sbmigWorkingDirectory}/datasources/`)
         await createJsonFile(
           JSON.stringify(res),
-          `${sbmigWorkingDirectory}/datasources/${filename}.json`
+          `${storyblokConfig.sbmigWorkingDirectory}/datasources/${filename}.json`
         )
         Logger.success(`All datasources written to a file:  ${filename}`)
       })
@@ -135,10 +136,10 @@ export default class Backup extends Command {
         if (res) {
           const randomDatestamp = new Date().toJSON()
           const filename = `datasource-${flags.oneDatasource}-${randomDatestamp}`
-          await createDir(`${sbmigWorkingDirectory}/datasources/`)
+          await createDir(`${storyblokConfig.sbmigWorkingDirectory}/datasources/`)
           await createJsonFile(
             JSON.stringify(res),
-            `${sbmigWorkingDirectory}/datasources/${filename}.json`
+            `${storyblokConfig.sbmigWorkingDirectory}/datasources/${filename}.json`
           )
           Logger.success(
             `Datasource for ${flags.oneDatasource} written to a file:  ${filename}`
@@ -156,10 +157,10 @@ export default class Backup extends Command {
         if (res) {
           const randomDatestamp = new Date().toJSON()
           const filename = `datasource-entries-${flags.datasourceEntries}-${randomDatestamp}`
-          await createDir(`${sbmigWorkingDirectory}/datasources/`)
+          await createDir(`${storyblokConfig.sbmigWorkingDirectory}/datasources/`)
           await createJsonFile(
             JSON.stringify(res),
-            `${sbmigWorkingDirectory}/datasources/${filename}.json`
+            `${storyblokConfig.sbmigWorkingDirectory}/datasources/${filename}.json`
           )
           Logger.success(
             `Datasource entries for ${flags.datasourceEntries} written to a file:  ${filename}`
@@ -177,10 +178,10 @@ export default class Backup extends Command {
         if (res) {
           const randomDatestamp = new Date().toJSON()
           const filename = `component-${flags.oneComponentPresets}-all_presets-${randomDatestamp}`
-          await createDir(`${sbmigWorkingDirectory}/component-presets/`)
+          await createDir(`${storyblokConfig.sbmigWorkingDirectory}/component-presets/`)
           await createJsonFile(
             JSON.stringify(res),
-            `${sbmigWorkingDirectory}/component-presets/${filename}.json`
+            `${storyblokConfig.sbmigWorkingDirectory}/component-presets/${filename}.json`
           )
           Logger.success(
             `Presets for ${flags.oneComponentPresets} written to a file:  ${filename}`
@@ -197,10 +198,10 @@ export default class Backup extends Command {
       return getAllPresets().then(async (res: any) => {
         const randomDatestamp = new Date().toJSON()
         const filename = `all-presets-${randomDatestamp}`
-        await createDir(`${sbmigWorkingDirectory}/presets/`)
+        await createDir(`${storyblokConfig.sbmigWorkingDirectory}/presets/`)
         await createJsonFile(
           JSON.stringify(res),
-          `${sbmigWorkingDirectory}/presets/${filename}.json`
+          `${storyblokConfig.sbmigWorkingDirectory}/presets/${filename}.json`
         )
         Logger.success(`All presets written to a file:  ${filename}`)
       })
@@ -215,10 +216,10 @@ export default class Backup extends Command {
         if (res) {
           const randomDatestamp = new Date().toJSON()
           const filename = `preset-${flags.onePreset}-${randomDatestamp}`
-          await createDir(`${sbmigWorkingDirectory}/presets/`)
+          await createDir(`${storyblokConfig.sbmigWorkingDirectory}/presets/`)
           await createJsonFile(
             JSON.stringify(res),
-            `${sbmigWorkingDirectory}/presets/${filename}.json`
+            `${storyblokConfig.sbmigWorkingDirectory}/presets/${filename}.json`
           )
           Logger.success(
             `Preset for '${flags.onePreset}' have been written to a file:  ${filename}`
