+++ src/commands/debug.ts
@@ -1,20 +1,26 @@
-import {Command, flags} from '@oclif/command'
-const configValues = require("../config/config")
-
+import { flags } from '@oclif/command'
+import Command from '../core'
+import {getAllComponents} from '../api/components'
 
 // TODO: implement --verbose flag to be available in every command
 export default class Debug extends Command {
   static description = 'Output extra debugging'
 
   static flags = {
-    help: flags.help({char: 'h'}),
+    help: flags.help({ char: 'h' }),
   }
 
   static args = []
 
   async run() {
-    const {args, flags} = this.parse(Debug)
+    const { args, flags } = this.parse(Debug)
 
-    this.log(configValues)
+    // this.getStoryblokConfig();
+    const config = this.getStoryblokConfig();
+
+    console.log("find components with ext");
+    console.log(this.findComponentsWithExt(config.schemaFileExt))
+    const dupa = await getAllComponents();
+    console.log(dupa)
   }
 }
